split-utils * Move useThemeColor to hooks/useThemeColor.ts
