+++ src/hooks/useThemeColor.ts
@@ -1,0 +1,14 @@
+import { useSettings } from '../settings/SettingsProvider'
+
+const useThemeColor = (
+    color: { [key: string]: string },
+    fallback: string,
+): string => {
+    const [settings] = useSettings()
+    if (!settings?.theme) {
+        return fallback
+    }
+    return color[settings?.theme] || fallback
+}
+
+export { useThemeColor }
